% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{create_forms}
\alias{create_forms}
\title{Creates CSAHS Misconduct forms}
\usage{
create_forms(
  usernames,
  assignment,
  assignment_grade,
  offense,
  assignment_due_date,
  assignment_submit_date,
  other_assignments,
  course,
  instructor,
  notes = "",
  interview_date = "today",
  forward_date = "today",
  chair,
  courselink_filename
)
}
\arguments{
\item{usernames}{Usernames for the students involved.}

\item{assignment}{Assignment name in Courselink gradebook.}

\item{assignment_grade}{Assignment grade (unofficial grade).}

\item{offense}{Academic offense. Use short form code(s).}

\item{assignment_due_date}{Assignment due date.}

\item{assignment_submit_date}{Assignment submission date.}

\item{other_assignments}{Other assignment grades (in Courselink gradebook) to be reported.}

\item{course}{Course code, name, and section number.}

\item{instructor}{Instructor's name.}

\item{notes}{Other notes about evidence from the instructor, long form text.}

\item{interview_date}{Date of form forwarded to Chair. If not specified, uses today's date.}

\item{chair}{Name of the Department Chair.}

\item{courselink_filename}{Name of the .csv grade file exported from Courselink.}
}
\description{
Creates CSAHS Misconduct forms
}
