% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misconduct.R
\name{format_people}
\alias{format_people}
\title{Format first and last name}
\usage{
format_people(people, remove_odd = FALSE)
}
\arguments{
\item{people}{A character vector of people's names}

\item{remove_odd}{If TRUE, delete names that might have been errors: NA, "A.", etc.}
}
\value{
A data.frame with columns first_name and last_name
}
\description{
Uses the humaniform package to get the first and last name.
}
